删除提示框

+++ src/app/pub-service/security-publicity/security-list/security-list.component.ts
@@ -10,15 +10,11 @@ import { MessageService } from 'primeng/components/common/messageservice';
 })
 export class SecurityListComponent implements OnInit {
 
-  // // 分页参数
-  // statisticPageNumber = 1;
-  // statisticTotal = 0;
-  // statisticFirst = 0;
-  // pageSize = 20;
-
   securityList: any = [];
   title: any;
+  articleId: any;
   detailVisible = false;
+  deleteDetailVisible = false;
 
   constructor(
     private SecurityListService: SecurityListService,
@@ -34,7 +30,6 @@ export class SecurityListComponent implements OnInit {
       .then(data => {
         if (data.status === 0) {
           this.securityList = data.data.list;
-          console.log(this.securityList);
         } else {
           this.messageService.add({ severity: 'warn', summary: '没有数据', detail: data.msg });
           this.securityList = [];
@@ -42,29 +37,27 @@ export class SecurityListComponent implements OnInit {
       });
   }
   delete = (articleId) => {
+    this.deleteDetailVisible = true;
+    this.articleId = articleId;
+  }
+  confirmDelete = () => {
     this.SecurityListService.deleteSecurityPublicityArticle({
-      articleId: articleId.articleId
+      articleId: this.articleId.articleId
     }).then(
       data => {
-        this.listSecurityPublicityArticle();
-      }
-    ).catch(
-      data => {
-        this.messageService.add({ severity: 'warn', summary: '获取登记统计信息失败', detail: data.msg });
+        if (data.status === 0) {
+          this.messageService.add({ severity: 'success', summary: '成功', detail: data.msg });
+          this.listSecurityPublicityArticle();
+        }else{
+          this.messageService.add({ severity: 'warn', summary: '获取登记统计信息失败', detail: data.msg });
+        }
+        this.deleteDetailVisible = false;
       }
     );
   }
   show = (data) => {
-    console.dir(data.articleDescription);
-    // document.getElementById('article').innerHTML = data.articleDescription;
     this.title = data.articleDescription;
     this.detailVisible = true;
   }
-  // onStatisticPageChange = (event) => {
-  //   this.statisticPageNumber = event.first / event.rows + 1;
-  //   this.pageSize = event.rows;
-  //   console.log(event);
-  //   // this.pageSize = pageSize;
-  // }
 
 }
